requirements.txt
# requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
sqlalchemy==2.0.23
alembic==1.12.1
psycopg[binary]==3.2.2
pydantic-settings==2.1.0
pydantic>=2.5.3
python-dotenv==1.0.0
email-validator==2.1.0.post1
cryptography==41.0.7